requirements.txt
fastapi>=0.109.1
uvicorn>=0.27.0
pydantic>=2.5.3
aiohttp>=3.10.2
loguru>=0.7.2
python-dotenv>=1.0.0
openai>=1.6.1
pyyaml>=6.0
websockets>=12.0
aiofiles>=23.0.0
Pillow>=10.0.0
requests>=2.31.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
asyncpg>=0.29.0
jieba>=0.42.1
tzdata>=2023.3
pytz>=2023.3
PyJWT>=2.8.0
passlib>=1.7.4
python-multipart>=0.0.6
dashscope>=1.23.9
cryptography>=42.0.0
aiortc>=1.9.0
av>=12.0.0
numpy>=1.24.0
